Created testing for ClientController.

+++ backend/src/main/java/br/com/uptimeowlet/backend/records/ClientInput.java
@@ -1,6 +1,6 @@
 package br.com.uptimeowlet.backend.records;
 
 public record ClientInput(String url, String method, String name,
-                          int httpCodeForCheckIfServiceIsActive, int checkPeriod, int timeoutConnection,
-                          int maxFailureForCheckIfServiceIsInactive, int periodForNewCheckAfterFailure) {
+                          int httpCodeForCheckIfServiceIsActive, long checkPeriod, long timeoutConnection,
+                          int maxFailureForCheckIfServiceIsInactive, long periodForNewCheckAfterFailure) {
 }
